requirements.txt
colorama~=0.4.3
discord.py